requirements.txt
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.1.1
Flask-Mail==0.9.1
Flask-Caching==2.0.1
SQLAlchemy==2.0.17
Werkzeug==2.2.2
python-dotenv==0.21.0
celery==5.2.7
redis==4.5.1
crontab==0.5.0